re #2818 Remove from MDDataObjects/MDDataAlgorithms under windows. some use of static_casts required. Should refactor out later.

+++ Code/Mantid/Framework/MDAlgorithms/src/CompositeImplicitFunction.cpp
@@ -61,7 +61,7 @@ namespace Mantid
          * */
         int CompositeImplicitFunction::getNFunctions() const
         {
-            return this->m_Functions.size();
+            return static_cast<int>( this->m_Functions.size() );
         }
